change error message

+++ db/connection.go
@@ -51,17 +51,22 @@ func (ldb *LegatoDB) GetUserConnectionById(u *User, id uint) (Connection, error)
 	return con, nil
 }
 
-func (ldb *LegatoDB) UpdateUserConnectionByID(u *User, name string, id uint) error {
+func (ldb *LegatoDB) UpdateUserConnectionNameByID(u *User, name string, id uint) error {
 	var connection Connection
 
 	err := ldb.db.Take(&Connection{}).
 		Where("ID = ?", id).Find(&connection).Error
 
-	if err == nil && connection.UserID == u.ID {
-		connection.Name = name
-		ldb.db.Session(&gorm.Session{FullSaveAssociations: true}).Updates(&connection)
-		ldb.db.Model(&connection).Update(connection.Token, name)
-		ldb.db.Save(&connection)
+	if err == nil {
+		if connection.UserID == u.ID {
+			connection.Name = name
+			ldb.db.Session(&gorm.Session{FullSaveAssociations: true}).Updates(&connection)
+			ldb.db.Model(&connection).Update(connection.Token, name)
+			ldb.db.Save(&connection)
+		} else {
+			return fmt.Errorf("there is no connection with this id for this user")
+		}
+
 	}
 	return err
 }
@@ -78,13 +83,16 @@ func (ldb *LegatoDB) CheckConnectionByID(u *User, id uint) error {
 	return fmt.Errorf("there is no connection with this id for this user")
 }
 
-func (ldb *LegatoDB) DeleteConnectionByID(id uint) error {
-	// ldb.db.Delete(&Connection{}, id)
-	// err := ldb.db.Take(&Connection{}).Where("id = ?", id).Delete(&Connection{}).Error
+func (ldb *LegatoDB) DeleteConnectionByID(u *User, id uint) error {
 	var connection Connection
 	err := ldb.db.Take(&Connection{}).
 		Where("ID = ?", id).Find(&connection).Error
-	ldb.db.Delete(&connection)
+	if u.ID == connection.UserID {
+		ldb.db.Delete(&connection)
+	} else {
+		return fmt.Errorf("there is not a connection with this id for this user")
+	}
+
 	return err
 }
 
@@ -94,11 +102,16 @@ func (ldb *LegatoDB) UpdateTokenFieldByID(u *User, Token string, id uint) error 
 	err := ldb.db.Take(&Connection{}).
 		Where("ID = ?", id).Find(&connection).Error
 
-	if err == nil && connection.UserID == u.ID {
-		connection.Token = Token
-		ldb.db.Session(&gorm.Session{FullSaveAssociations: true}).Updates(&connection)
-		ldb.db.Model(&connection).Update(connection.Token, Token)
-		ldb.db.Save(&connection)
+	if err == nil {
+		if connection.UserID == u.ID {
+			connection.Token = Token
+			ldb.db.Session(&gorm.Session{FullSaveAssociations: true}).Updates(&connection)
+			ldb.db.Model(&connection).Update(connection.Token, Token)
+			ldb.db.Save(&connection)
+		} else {
+			return fmt.Errorf("there is no connection with this id for this user")
+		}
+
 	}
 	return err
 }
